+++ packages/core/src/hooks/translate/useGetLocale.ts
@@ -1,7 +1,7 @@
 import { useContext, useCallback } from "react";
 import { TranslationContext } from "@contexts/translation";
 
-export type UseGetLocaleType = () => () => string;
+export type UseGetLocaleType = () => () => string | undefined;
 
 /**
  * If you need to know the current locale, refine provides the `useGetLocale` hook.
@@ -12,5 +12,5 @@ export type UseGetLocaleType = () => () => string;
 export const useGetLocale: UseGetLocaleType = () => {
     const { i18nProvider } = useContext(TranslationContext);
 
-    return useCallback(() => i18nProvider.getLocale(), []);
+    return useCallback(() => i18nProvider?.getLocale(), []);
 };