+++ packages/antd/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @pankod/refine-antd
 
+## 3.23.2
+
+### Patch Changes
+
+-   [#1873](https://github.com/pankod/refine/pull/1873) [`2deb19babf`](https://github.com/pankod/refine/commit/2deb19babfc6db5b00b111ec29aa5ece4c371bbc) Thanks [@aliemir](https://github.com/aliemir)! - Removed dummy default values from internal contexts.
+    Updated contexts:
+
+    -   Auth
+    -   Access Control
+    -   Notification
+    -   Translation (i18n)
+    -   unsavedWarn
+
+    **BREAKING:** `useGetLocale` hook now can return `undefined` instead of a fallback value of `en` in cases of `i18nProvider` being `undefined`.
+
+-   Updated dependencies [[`2deb19babf`](https://github.com/pankod/refine/commit/2deb19babfc6db5b00b111ec29aa5ece4c371bbc)]:
+    -   @pankod/refine-core@3.23.2
+
 ## 3.23.1
 
 ### Patch Changes